requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic-settings==2.7.1

# ── LangChain Ecosystem ────────────────────────────────────────────────────────
langchain==0.3.14
langchain-community==0.3.14
langchain-mongodb==0.3.0

# ── Local Embeddings (no API key — runs 100% on your machine) ─────────────────
# sentence-transformers downloads the model once (~500 MB) on first run,
# then works fully offline.  Model used: paraphrase-multilingual-mpnet-base-v2
# (768-dim, supports Vietnamese and 50+ languages).
sentence-transformers==3.4.1

# ── Document Parsers ──────────────────────────────────────────────────────────
pypdf==5.1.0
docx2txt==0.8

# ── Database ──────────────────────────────────────────────────────────────────
pymongo==4.10.1

# ── Telegram Bot ──────────────────────────────────────────────────────────────
python-telegram-bot==22.7
httpx==0.28.1            # async HTTP client — used to download Telegram files

# ── Anthropic (Claude Vision + RAG answer generation) ─────────────────────────
anthropic==0.49.0

# ── Environment & Utilities ───────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.4.2
pytest-asyncio==0.25.3
pytest-httpx==0.35.0
